pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tool-orchestra"
version = "0.1.0"
description = "Cost-aware multi-model orchestration framework based on ToolOrchestra paper"
readme = "PLAN.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["llm", "orchestration", "langgraph", "agents", "tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # LLM Clients
    "openai>=1.0.0",       # LM Studio (OpenAI-compatible API)
    "google-genai>=1.0.0", # Gemini API (unified GenAI SDK)

    # Vector Store & Embeddings
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
    "chromadb>=0.4.0; sys_platform != 'darwin' or platform_machine != 'arm64'", # Issues on M1/M2/M3

    # Core utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Async & HTTP
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",

    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",

    # Caching
    "diskcache>=5.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
api = ["fastapi>=0.109.0", "uvicorn[standard]>=0.27.0"]
langsmith = ["langsmith>=0.1.0"]
all = ["tool-orchestra[dev,api,langsmith]"]

[project.scripts]
orchestra = "src.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/tool-orchestra"
Documentation = "https://github.com/yourusername/tool-orchestra#readme"
Repository = "https://github.com/yourusername/tool-orchestra"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

# Tool subclasses intentionally extend run()/arun() signatures
# with specific parameters (Liskov substitution trade-off for usability)
[[tool.mypy.overrides]]
module = "src.tools.*"
disable_error_code = ["override"]

[[tool.mypy.overrides]]
module = "src.orchestrator.graph"
disable_error_code = ["type-var"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
    "e2e: marks end-to-end tests",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
